Many fixes
Changed code from routes.py into routesUser.py, added error handling, finished user delete, fixes bugs and fixed registration
