+++ myapp/forms.py
@@ -5,6 +5,11 @@ from wtforms.validators import DataRequired
 class LoginForm(FlaskForm):
     username = StringField('Username', validators=[DataRequired()])
     password = PasswordField('Password')
+    remember_me = BooleanField('Remember me')
     submit = SubmitField('Sign in')
 
-
+class RegisterForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired()])
+    password = PasswordField('Password', validators=[DataRequired()])
+    repeatPassword = PasswordField('Repeat Password', validators=[DataRequired()])
+    submit = SubmitField('Register')
