+++ myapp/templates/register.html
@@ -12,6 +12,9 @@
     <p> {{ form.password.label }} <br>
     {{ form.password(size=32) }}</p>
 
+    <p> {{ form.repeatPassword.label }} <br>
+    {{ form.repeatPassword(size=32) }}</p>
+
     <p>{{ form.submit() }}</p>
 </form>
 {% endblock %}
